fhir-preprocessor/src/requirements.txt
flask==1.1.1
flask-restful-swagger-2==0.35
flask-cors==3.0.8
flask_sqlalchemy==2.4.1
psycopg2-binary==2.8.4
requests==2.23.0
cerberus==1.3.2
uwsgi==2.0.18
eventlet==0.25.1
pymongo==3.10.1
sqlparse==0.3.1